+++ packages/omgx/wallet-frontend/src/containers/transactions/Transactions.js
@@ -14,225 +14,70 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 
 import React, { useState, useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { isEqual, orderBy } from 'lodash';
 import { useSelector } from 'react-redux';
-import DatePicker from 'react-datepicker';
 import "react-datepicker/dist/react-datepicker.css";
-
 
 import moment from 'moment';
 import truncate from 'truncate-middle';
 
-import { setActiveHistoryTab1 } from 'actions/uiAction'
-import { setActiveHistoryTab2 } from 'actions/uiAction'
-
-import { selectActiveHistoryTab1 } from 'selectors/uiSelector'
-import { selectActiveHistoryTab2 } from 'selectors/uiSelector'
-
-import { selectTransactions } from 'selectors/transactionSelector';
 import { selectLoading } from 'selectors/loadingSelector'
 
-import Tabs from 'components/tabs/Tabs'
-import Input from 'components/input/Input'
 import Transaction from 'components/transaction/Transaction'
 import Pager from 'components/pager/Pager'
 
-import Exits from './Exits';
-import Deposits from './Deposits';
-
 import * as styles from './Transactions.module.scss';
 
-import { getAllNetworks } from 'util/masterConfig'
-import { selectNetwork } from 'selectors/setupSelector'
+const PER_PAGE = 10;
 
-const PER_PAGE = 8;
-
-function Transactions () {
-
-  const dispatch = useDispatch();
-
-  const [ page1, setPage1 ] = useState(1);
-  // eslint-disable-next-line
-  const [ page2, setPage2 ] = useState(1);
-
-  const start = new Date();
-  const end = new Date();
-  const [startDate, setStartDate] = useState(new Date(start.setDate(1)));
-  const [endDate, setEndDate] = useState(new Date(end.setDate(end.getDate() + 7)));
-
-  const [ searchHistory, setSearchHistory ] = useState('');
-
-  const loading = useSelector(selectLoading([ 'TRANSACTION/GETALL' ]));
-
-  const activeTab1 = useSelector(selectActiveHistoryTab1, isEqual);
-  const activeTab2 = useSelector(selectActiveHistoryTab2, isEqual);
-
-  const unorderedTransactions = useSelector(selectTransactions, isEqual)
-
-  // const transactions = orderBy(unorderedTransactions, i => i.timeStamp, 'desc');
-  const [transactions,setTransactions] = useState([]);
-  const orderedTransactions = orderBy(unorderedTransactions, i => i.timeStamp, 'desc');
+function Transactions({ searchHistory, transactions, chainLink }) {
+  const [page, setPage] = useState(1);
+  const loading = useSelector(selectLoading(['EXIT/GETALL']));
+  useEffect(() => {
+    setPage(1);
+  }, [searchHistory]);
 
   const _transactions = transactions.filter(i => {
     return i.hash.includes(searchHistory);
   });
 
-  const startingIndex = page1 === 1 ? 0 : ((page1 - 1) * PER_PAGE);
-  const endingIndex = page1 * PER_PAGE;
+  const startingIndex = page === 1 ? 0 : ((page - 1) * PER_PAGE);
+  const endingIndex = page * PER_PAGE;
   const paginatedTransactions = _transactions.slice(startingIndex, endingIndex);
 
   let totalNumberOfPages = Math.ceil(_transactions.length / PER_PAGE);
 
   //if totalNumberOfPages === 0, set to one so we don't get the strange "page 1 of 0" display
-  if (totalNumberOfPages === 0) totalNumberOfPages = 1;
+  if (totalNumberOfPages === 0) totalNumberOfPages = 1
 
-  const currentNetwork = useSelector(selectNetwork());
-
-  const nw = getAllNetworks();
-
-  const chainLink = (item) => {
-    let network = nw[currentNetwork];
-    if (!!network && !!network[item.chain]) {
-      // network object should have L1 & L2
-      return `${network[item.chain].transaction}${item.hash}`;
-    }
-    return '';
-  }
-
-  useEffect(()=>{
-    if(orderedTransactions) {
-      const filtered = orderedTransactions.filter((i)=>{
-        return (moment.unix(i.timeStamp).isAfter(startDate) && moment.unix(i.timeStamp).isBefore(endDate));
-      })
-      setTransactions(filtered);
-      setPage1(1)
-    }
-  },[startDate, endDate]);
-  
-  useEffect(()=>{
-    if(orderedTransactions) {
-      const filtered = orderedTransactions.filter((i)=>{
-        return (moment.unix(i.timeStamp).isAfter(startDate) && moment.unix(i.timeStamp).isBefore(endDate));
-      })
-      setTransactions(filtered);
-      setPage1(1)
-    }
-  },[]);
-
-  return (
-    <div className={styles.container}>
-
-      <div className={styles.header}>
-        <h2>Search</h2>
-        <div className={styles.actions}>
-          <div>Start: </div>
-          <DatePicker
-            selected={startDate}
-            onChange={(date) => setStartDate(date)}
-            selectsStart
-            startDate={startDate}
-            endDate={endDate}
-          />
-
-          <div>End: </div>
-          <DatePicker
-            selected={endDate}
-            onChange={(date) => setEndDate(date)}
-            selectsEnd
-            startDate={startDate}
-            endDate={endDate}
-            minDate={startDate}
-          />
-          <Input
-            icon
-            placeholder='Search by hash'
-            value={searchHistory}
-            onChange={i => {
-              setPage1(1);
-              setSearchHistory(i.target.value);
-            }}
-            className={styles.searchBar}
-          />
-        </div>
-      </div>
-      <div className={styles.data}>
-
-        <div className={styles.section}>
-        
-          <Tabs
-            onClick={tab => {
-              setPage1(1);
-              dispatch(setActiveHistoryTab1(tab));
-            }}
-            activeTab={activeTab1}
-            tabs={[ 'Transactions', 'Deposits' ]}
-          />
-
-          {activeTab1 === 'Transactions' && (
-            <div className={styles.transactions}>
-              <Pager
-                currentPage={page1}
-                isLastPage={paginatedTransactions.length < PER_PAGE}
-                totalPages={totalNumberOfPages}
-                onClickNext={()=>setPage1(page1 + 1)}
-                onClickBack={()=>setPage1(page1 - 1)}
-              />
-              {!paginatedTransactions.length && !loading && (
-                <div className={styles.disclaimer}>Transaction history coming soon...</div>
-              )}
-              {!paginatedTransactions.length && loading && (
-                <div className={styles.disclaimer}>Loading...</div>
-              )}
-              {paginatedTransactions.map((i, index) => {
-                const metaData = typeof(i.typeTX) === 'undefined' ? '' : i.typeTX
-                return (
-                  <Transaction
-                    key={index}
-                    link={chainLink(i)}
-                    title={`${truncate(i.hash, 8, 6, '...')}`}
-                    midTitle={moment.unix(i.timeStamp).format('lll')}
-                    blockNumber={`Block ${i.blockNumber}`}
-                    chain={`${i.chain} Chain`}
-                    typeTX={`${metaData}`}
-                  />
-                )
-              })}
-            </div>
-          )}
-
-          {activeTab1=== 'Deposits' && <
-            Deposits 
-              searchHistory={searchHistory} 
-              transactions={transactions} 
-            />
-          }
-
-        </div>
-
-        <div className={styles.section}>
-          <Tabs
-            onClick={tab => {
-              setPage2(1);
-              dispatch(setActiveHistoryTab2(tab));
-            }}
-            activeTab={activeTab2}
-            tabs={[ 'Exits', 'TBD' ]}
-          />
-
-          {activeTab2 === 'Exits' && 
-            <Exits 
-              searchHistory={searchHistory}
-              transactions={transactions}  
-              chainLink={chainLink}
-            />
-          }
-
-        </div>
-
-      </div>
-    </div>
-  );
+  return (<div className={styles.transactions}>
+    <Pager
+      currentPage={page}
+      isLastPage={paginatedTransactions.length < PER_PAGE}
+      totalPages={totalNumberOfPages}
+      onClickNext={() => setPage(page + 1)}
+      onClickBack={() => setPage(page - 1)}
+    />
+    {!paginatedTransactions.length && !loading && (
+      <div className={styles.disclaimer}>Transaction history coming soon...</div>
+    )}
+    {!paginatedTransactions.length && loading && (
+      <div className={styles.disclaimer}>Loading...</div>
+    )}
+    {paginatedTransactions.map((i, index) => {
+      const metaData = typeof (i.typeTX) === 'undefined' ? '' : i.typeTX
+      return (
+        <Transaction
+          key={index}
+          link={chainLink(i)}
+          title={`${truncate(i.hash, 8, 6, '...')}`}
+          midTitle={moment.unix(i.timeStamp).format('lll')}
+          blockNumber={`Block ${i.blockNumber}`}
+          chain={`${i.chain} Chain`}
+          typeTX={`${metaData}`}
+        />
+      )
+    })}
+  </div>)
 }
 
 export default React.memo(Transactions);
